Reorganize ExecutionUtils in FirrtlExecutionUtils
